Add the (to be) inserted or (to be) removed html to the event.
The event gets an extra parameter: for `before-insert` and `after-insert` the inserted piece of html. For `before-remove` and `after-remove` the removed piece of html. This way one can add effects/animations if needed.

The remove code has been refactored (less code duplication), and checks for a possible timeout, which can be set in the `before-remove` callback.