Add multiple items on one click
I've added the option of a `count` parameter in #link_to_add_association helper.
When this is a valid integer, every click on the Add button will create that number of items.

Most of the code is in the javascript of course.
I also cleaned the code a little (newcontent_braced, newcontent_underscord function and create_new_id functions)

In my use case, next to the Add button I added a select tag that changes the count settings.

Example:

Markup (Haml):
  .add-task-items
    = link_to_add_association 'Add', f, :association
    %select
      %option 1
      %option 3
      %option 5

Code (JS):
  $('.add-task-items').on('change', 'select', function({
    count = $(this).val();
    $(this).closest('.add-task-items').find('.add_fields').data('count', count);
  });