Removed Jeweler and created modern .gemspec.
The cocoon.gemspec generate by Jeweler was including the rubysl gems. This
causes cocoon to fail to install on MRI right now (see eg this stackoverflow
post: http://stackoverflow.com/questions/20390594/rails-gem-install-error-coccon)

Jeweler was also emitting redundant dependencies for rspec and emitting a
dependency for actionpack despite that the dependency on rails would pull
in actionpack.

To replace the Jeweler-generated cocoon.gemspec, I added one created by
Bundler and migrated the settings over. It is a simple matter to keep this
gemspec current by hand.