"Unfreeze" frozen objects.
In rails 4, if a `belongs_to` association is deleted,
the previous associated object is frozen as a
precaution to not keep using it by accident. But in
cocoon we generate the object on the association, 
delete it, _and_ keep using it on purpose.

So if the `assoc_obj` is frozen, create a copy
and use that instead.