Change symbol syntax in readme
Hello, I've decided to change symbol syntax to new one in readme. If you do not like new syntax for some reason, feel free to close this pull request. I am new to ruby and started to use it from version 2, so I am really get it better when I see symbol: 'something' and not :symbol => 'something'. As I can see, rails guys are moving to new syntax too ^_^

Readme is a place where you start to learn about cocoon, I think it should be up to date with Ruby :)))