prevents propagation to parent elements
So the problem I have is that I display a nested form in a dropdown (eg: list of items in a shopping cart). When I click on the ```.remove_fields.existing``` element the dropdown is closed because the click event propagates through the parent elements which is not the expected behavior and I also think that is good practice to prevent side-effects that may cause problems to the library users.